refactor: extract Socket.IO configuration into separate module #549
